server/requirements.txt
flask-restful
sqlalchemy
mysqlclient